Refactor landing folders
